[fix] context and loader design.
