+++ go/svc/graphql/pkg/loader/look_up.go
@@ -1,0 +1,17 @@
+package loader
+
+import (
+	"github.com/graph-gophers/dataloader"
+)
+
+type key string
+
+const (
+	userKey = "user"
+)
+
+var (
+	LookUpTable = map[key]dataloader.BatchFunc{
+		userKey: batchLoadUser,
+	}
+)
